+++ src/types/event-handler.js
@@ -1,86 +1,83 @@
 /**
  * # Eventable
  * Class for generic event handling.
- * Initial events are loaded as a generic object of ```<name>:<callback>```.
- * For example:
  * ```javascript
- * Eventable({
+ * EventHandler({
  *     events: {
  *         eventName: function () {},
- *         otherName: [
- *             function () {},
- *             function () {}
- *         ]
+ *         otherName: function () {}
  *     },
  *     singles: {
  *         anotherName: function () {}
  *     }
  * })
  * ```
- * @param {Object} [opts.events] Event callbacks to fire each time
+ * @constructor
+ * @param {Object<string, function(?)>} [opts.events] Event callbacks to fire each time
  * the event is triggered.
- * @param {Object} [opts.singles] Event callbacks to fire only once
+ * @param {Object<string, function(?)>} [opts.singles] Event callbacks to fire only once
  * the next time the event is triggered.
- * @return {Object} Can be used by BaseClass.extend().
+ * @return {EventHandler}
  */
 module.exports = function (opts) {
     var events = {},
         singles = {},
         name,
-        // Cache of events that have already been triggered once.
+        /**
+         * ## triggerCache
+         * Cache of events that have already been triggered once.
+         * @private
+         * @type {Object<string, !*>}
+         */
         triggerCache = {};
 
     // Load in the initial callbacks.
     opts = opts || {};
     for (name in opts.events) {
-        events[name] = [].concat(
-            opts.events[name] || []
-        );
+        events[name] = opts.events[name];
     }
     for (name in opts.singles) {
-        singles[name] = [].concat(
-            opts.singles[name] || []
-        );
+        singles[name] = opts.singles[name];
     }
 
     return {
         /**
-         * ## harmony.on(name, callback, [lazy])
+         * ## handler.on(name, callback, [lazy])
          * Bind a callback to an event. Callback will run each
          * time the event is triggered.
-         * @param {String} name Event name.
-         * @param {Function} cb Event callback.
-         * @param {Boolean} [lazy] True if this callback should not fire
+         * @param {string} name Event name.
+         * @param {function(?)} cb Event callback.
+         * @param {boolean} [lazy] True if this callback should not fire
          * immediately if the event has already been triggered.
          */
         on: function (name, cb, lazy) {
             if (!lazy && name in triggerCache) {
-                cb.call(this, triggerCache[name].data);
+                cb(triggerCache[name]);
             }
             events[name] = events[name] || [];
             events[name].push(cb);
         },
         /**
-         * ## harmony.one(name, callback, [lazy])
+         * ## handler.one(name, callback, [lazy])
          * Bind a callback to an event. Callback will run only
          * once the next time the event is triggered.
-         * @param {String} name Event name.
-         * @param {Function} cb Event callback.
-         * @param {Boolean} [lazy] True if this callback should not fire
+         * @param {string} name Event name.
+         * @param {function(?)} cb Event callback.
+         * @param {boolean} [lazy] True if this callback should not fire
          * immediately if the event has already been triggered.
          */
         one: function (name, cb, lazy) {
             if (!lazy && name in triggerCache) {
-                cb.call(this, triggerCache[name].data);
+                cb(triggerCache[name]);
             } else {
                 singles[name] = singles[name] || [];
                 singles[name].push(cb);
             }
         },
         /**
-         * ## harmony.off([name])
+         * ## handler.off([name])
          * Clear all callbacks from an event.
-         * @param {String} [name] Event name. Empty to
+         * @param {string} [name] Event name. Empty to
          * clear all callbacks from the system.
          */
         off: function (name) {
@@ -94,29 +91,26 @@ module.exports = function (opts) {
             }
         },
         /**
-         * ## harmony.trigger(name, data)
+         * ## handler.trigger(name, [data])
          * Immediately trigger an event.
-         * @param {String} name Event name.
-         * @param {Any} [data] Any data you wish to provide to all
+         * @param {string} name Event name.
+         * @param {*} [data] Any data you wish to provide to all
          * event callbacks.
          */
         trigger: function (name, data) {
-            var that = this;
             if (name in events) {
                 events[name].forEach(function (cb) {
-                    cb.call(that, data);
+                    cb(data);
                 });
             }
             if (name in singles) {
                 singles[name].forEach(function (cb) {
-                    cb.call(that, data);
+                    cb(data);
                 });
                 singles[name] = [];
             }
             // Cache that this event has fired.
-            triggerCache[name] = {
-                data: data
-            };
+            triggerCache[name] = data;
         }
     };
 };